fix: 라우트 에러 해결
- routes.tsx를 main.tsx 과 분리하니 에러가 발생하여 main.tsx에 합쳐서 해결
