+++ src/App.tsx
@@ -3,8 +3,8 @@ import {RouterProvider} from "react-router-dom";
 // 전역 상태
 import ResetStyle from "@/styles/ResetStyle.ts";
 import GlobalStyle from "./styles/GlobalStyle.ts";
+import {router} from "@/main.tsx";
 
-import {router} from "@/routes.tsx";
 
 function App() {
   return (
